Phase 9 #42: memory64 + table64, and a zero-partial spec suite
Enable WasmFeatures::MEMORY64 (covers both memory64 and table64). The memory
type layer was already wired; add the missing table64 flag to IrTableType
(+ u64 limits) and TableType (new64/is_64, read from tt.table64).

Core mechanic — "permissive widening pop": Execution::pop_index() pops
whichever int variant validation placed (i32 for 32-bit, i64 for 64-bit) and
zero-extends to u64, so no per-op index-width encoding is needed; only
size/grow *results* use push_index(is_64, ..). Widen MemArg.offset u32->u64
(mem_ea overflow-traps in u64); thread pop_index through every index-typed
operand (load/store addr, memory/table fill/copy/init dst+len, grow delta,
table.get/set, call_indirect index), keeping data/elem-segment src+len at i32.
checked_range (memory.rs + table.rs) now does u64 arithmetic before the usize
slice cast.

Security: index-width-dependent grow ceiling (memory64 -> 2^48 pages) plus
fallible Vec growth (try_reserve_exact -> None/-1) so an over-large guest grow
returns -1 instead of OOM-aborting; import matching compares the index type
(a 32<->64 mismatch is unlinkable). Artifact version 8->9.

Cap: relocate TagType to src/value/tag_type.rs to keep types.rs under 400.

Spec gate +5,994 asserts (30,782 -> 36,776): memory_copy64 (4417),
table_copy64 (1675), memory64, load64, address64, call_indirect64, etc. green.

Harness (tests/spec.rs) — close the last non-SIMD gaps so the suite is
0-partial and SIMD-only:
- arg_to_val builds null refs for the any-hierarchy (any/eq/i31/struct/array/
  none) and exn-hierarchy (exn/noexn) -> extern.wast fully green.
- handle the module-definition / module-instance Wasm-3.0 directives: a
  definition is validated+compiled but not instantiated (so a huge
  (memory 65536) definition allocates nothing); a module-instance instantiates
  a stored definition under a fresh name -> instance.wast fully green.
- spectest shim exports a 64-bit table64.

Result: 190 files fully ok, 0 partial; every remaining skip is SIMD/relaxed.
Integration test tests/memory64.rs.

+++ src/module/op.rs
@@ -19,7 +19,7 @@ pub(crate) type ElemIdx = u32;
 #[derive(Copy, Clone, Debug, serde::Serialize, serde::Deserialize)]
 pub(crate) struct MemArg {
     pub memory: u32,
-    pub offset: u32,
+    pub offset: u64,
 }
 
 /// A resolved branch edge: on a taken branch the top `keep` operands are moved down over
